# querystring: improve parse() performance