--- a/lib/querystring.js
+++ b/lib/querystring.js
@@ -159,6 +159,7 @@ QueryString.stringify = QueryString.encode = function(obj, sep, eq, options) {
   return '';
 };
 
+var RE_PLUS = /\+/g;
 // Parse a key=val string.
 QueryString.parse = QueryString.decode = function(qs, sep, eq, options) {
   sep = sep || '&';
@@ -169,7 +170,7 @@ QueryString.parse = QueryString.decode = function(qs, sep, eq, options) {
     return obj;
   }
 
-  var regexp = /\+/g;
+  var regexp = RE_PLUS;
   qs = qs.split(sep);
 
   var maxKeys = 1000;
@@ -201,13 +202,15 @@ QueryString.parse = QueryString.decode = function(qs, sep, eq, options) {
       vstr = '';
     }
 
-    try {
-      k = decode(kstr);
-      v = decode(vstr);
-    } catch (e) {
-      k = QueryString.unescape(kstr, true);
-      v = QueryString.unescape(vstr, true);
-    }
+    if (~kstr.indexOf('%'))
+      k = tryDecode(decode, kstr);
+    else
+      k = kstr;
+
+    if (~vstr.indexOf('%'))
+      v = tryDecode(decode, vstr);
+    else
+      v = vstr;
 
     if (!hasOwnProperty(obj, k)) {
       obj[k] = v;
@@ -220,3 +223,11 @@ QueryString.parse = QueryString.decode = function(qs, sep, eq, options) {
 
   return obj;
 };
+
+function tryDecode(fn, str) {
+  try {
+    return fn(str);
+  } catch (e) {
+    return QueryString.unescape(str, true);
+  }
+}