@jasnell 2015-08-15
@mscdex ... do you still want to pursue this on v0.12?